requirements.txt
appdirs==1.4.3
asgiref==3.2.5
astroid==2.3.3
beautifulsoup4==4.8.2
bootstrap4==0.1.0
certifi==2019.11.28
cloudinary==1.20.0
colorama==0.4.3
distlib==0.3.0
Django==3.0.3
django-cloudinary-storage==0.2.3
django-bootstrap4==1.1.1
django-forms-bootstrap==3.1.0
django-taggit==1.2.0
filelock==3.0.12
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
olefile==0.46
Pillow==7.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.5
pylint==2.4.4
pytz==2019.3
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
urllib3==1.25.8
virtualenv==20.0.15
whitenoise==5.0.1
wincertstore==0.2
wrapt==1.11.2
